Add cross-links to recipes pages
I have added an unordered list of links to the
other recipes on each recipe page at the bottom.
For example: the fried rice page how has a list
with links to the pasta and entomatadas recipes.

I have also added the target="_blank" and
rel="noopener noreferrer" attributes in each <a>
tag for safety reasons.